[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ C++20/백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -3,12 +3,12 @@
 using namespace std;
 
 
-int oper(int a, int b) {
+long long int oper(long long int a, long long int b) {
 	return (a + b) * (a - b);
 }
 
 int main() {
-	int a, b;
+	long long int a, b;
 
 	cin >> a >> b;
 	cout << oper(a, b) << endl;
